spi: introduce spi_delay struct as value + unit + spi_delay_exec()
There are plenty of delays that have been introduced in SPI core. Most of
them are in micro-seconds, some need to be in nano-seconds, and some in
hertz/clock-cycles.

For some of these delays (related to transfers & CS timing) it may make
sense to have a `spi_delay` struct that abstracts these a bit.

The important element of these delays [for unification] seems to be the
`unit` of the delay.

This change adds the `spi_delay` struct and makes the `cs_change_delay` the
first user of this. The intent is to add support for more delays on the
`spi_transfer` (such as CS setup-delay, hold-delay, inactive-delay, etc).

These can be specified during a transfer and the framework could handle
them. But before doing that, a unification/consensus on the delay format
should be reached, so that people would be allowed to specify delay units.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>